+++ src/main/java/domain/user/controller/UserController.java
@@ -1,25 +1,23 @@
 package domain.user.controller;
 
 import domain.user.service.UserService;
-import utils.RequestUtils;
 import webserver.FileResponse;
 import webserver.HttpStatus;
 import webserver.Request;
 import webserver.Response;
 
-import java.util.Map;
-
 public class UserController {
 
     public static Response getCreateUser(Request request) {
-        Map<String, String> inputs = RequestUtils.extractQueryString(request.getQueryString());
-        UserService.createUser(inputs);
+        String input = request.extractQueryString();
+        UserService.createUser(input);
         return Response.withFileResponse(HttpStatus.OK, new FileResponse("./templates/index.html", "text/html"));
     }
 
     public static Response postCreateUser(Request request) {
-        Map<String, String> inputs = RequestUtils.extractQueryString(request.getBody());
-        UserService.createUser(inputs);
+        String input = request.extractBody();
+        UserService.createUser(input);
         return Response.withLocation(HttpStatus.FOUND, "/index.html");
+
     }
 }
