+++ src/main/java/domain/user/dto/UserRequest.java
@@ -2,16 +2,29 @@ package domain.user.dto;
 
 import domain.user.User;
 
+import java.util.Arrays;
+import java.util.List;
+import java.util.Objects;
+
 public class UserRequest {
     private static final String USER_ID = "userId";
     private static final String PASSWORD = "password";
     private static final String NAME = "name";
     private static final String EMAIL = "email";
+    private static final String USER_REQUEST_REGEX = "userId=([^&=]+)&password=([^&=]+)&name=([^&=]+)&email=([^&=]+)";
 
     private String userId;
     private String password;
     private String name;
     private String email;
+
+    public List<String> getFieldNames() {
+        return Arrays.asList(USER_ID, PASSWORD, NAME, EMAIL);
+    }
+
+    public boolean isLastField(String fieldName) {
+        return Objects.equals(fieldName, EMAIL);
+    }
 
     public void set(String fieldName, String value) {
         if (USER_ID.equals(fieldName)) {
@@ -30,5 +43,9 @@ public class UserRequest {
 
     public User toUser() {
         return new User(this.userId, this.password, this.name, this.email);
+    }
+
+    public String getRegex() {
+        return USER_REQUEST_REGEX;
     }
 }
