+++ src/main/java/domain/user/service/UserService.java
@@ -2,15 +2,12 @@ package domain.user.service;
 
 import db.DataBase;
 import domain.user.User;
-import utils.UserMapper;
-
-import java.util.Map;
+import utils.RequestUtils;
 
 public class UserService {
 
-    public static void createUser(Map<String, String> inputs) {
-        User user = UserMapper.createUser(inputs);
+    public static void createUser(String input) {
+        User user = RequestUtils.createUser(input);
         DataBase.addUser(user);
-        System.out.println(DataBase.findAll());
     }
 }
